src/tools_com/com_Material/com_display.py
"""
Module com_display - Fonctions de présence, fonctionnement et utilisation de l'affichage
"""

import logging

logger = logging.getLogger(__name__)

def check_display_presence():
    """
    Vérifie la présence de l'affichage.

    :return: True si l'affichage est présent, False sinon.
    :rtype: bool
    """
    try:
        # Vérification de la présence de l'affichage ici
        presence = True  # Exemple de réponse (à adapter)
        if presence:
            logger.info("Display is present.")
        else:
            logger.info("Display is not present.")
        return presence
    except Exception:
        logger.error("Failed to check display presence.")
        return False

def start_display():
    """
    Démarre l'affichage pour l'affichage visuel.

    :return: True si l'affichage a démarré avec succès, False sinon.
    :rtype: bool
    """
    try:
        # Démarrage de l'affichage ici
        started = True  # Exemple de réponse (à adapter)
        if started:
            logger.info("Display started.")
        else:
            logger.error("Failed to start display.")
        return started
    except Exception:
        logger.error("Failed to start display.")
        return False

def show_text(text):
    """
    Affiche du texte sur l'affichage.

    :param text: Texte à afficher.
    :type text: str
    :return: True si le texte a été affiché avec succès, False sinon.
    :rtype: bool
    """
    try:
        # Affichage du texte sur l'affichage
        # Utilisation du texte spécifié par le paramètre text
        displayed = True  # Exemple de réponse (à adapter)
        if displayed:
            logger.info("Text displayed on the display.")
        else:
            logger.error("Failed to display text.")
        return displayed
    except Exception:
        logger.error("Failed to display text.")
        return False

def clear_display():
    """
    Efface le contenu de l'affichage.

    :return: True si le contenu de l'affichage a été effacé avec succès, False sinon.
    :rtype: bool
    """
    try:
        # Effacement du contenu de l'affichage ici
        cleared = True  # Exemple de réponse (à adapter)
        if cleared:
            logger.info("Display content cleared.")
        else:
            logger.error("Failed to clear display content.")
        return cleared
    except Exception:
        logger.error("Failed to clear display content.")
        return False
